requirements.txt
astroid==2.4.2
autopep8==1.5.3
click==7.1.2
Flask==1.1.2
Flask-SocketIO==4.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
obs-websocket-py==0.5.1
pycodestyle==2.6.0
pylint==2.5.3
python-engineio==3.13.0
python-socketio==4.6.0
six==1.15.0
toml==0.10.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
